Move OAuth helpers into utils.php
Add a new OAuth\Utils file with create_state, create_return_to and ba_alert, centralizing repeated helper logic. Update callback.php, login.php and oauth/user_infos.php to import and use these utilities (simplifying state building and callback URL creation, and removing duplicated functions). Note: create_return_to now accepts the HTTP referer as an argument and create_state consolidates GET parameter sanitization.
