+++ src/login.php
@@ -3,6 +3,8 @@
 use MediaWiki\OAuthClient\Client;
 use MediaWiki\OAuthClient\ClientConfig;
 use MediaWiki\OAuthClient\Consumer;
+use function OAuth\Utils\create_state;
+use function OAuth\Utils\create_return_to;
 
 $env = getenv('APP_ENV') ?: ($_ENV['APP_ENV'] ?? 'development');
 
@@ -62,25 +64,6 @@ try {
     showErrorAndExit("An internal error occurred while preparing the authentication service. Please try again later.");
 }
 
-function create_return_to()
-{
-
-    $allowed_domains = ['mdwiki.toolforge.org', 'localhost'];
-    $return_to = '';
-    if (isset($_SERVER['HTTP_REFERER'])) {
-        $parsed = parse_url($_SERVER['HTTP_REFERER']);
-        if (isset($parsed['host']) && in_array($parsed['host'], $allowed_domains)) {
-            $return_to = $_SERVER['HTTP_REFERER'];
-        }
-    }
-
-    if (!empty($return_to) && (strpos($return_to, '/auth/') !== false)) {
-        $return_to = "";
-    }
-
-    return $return_to;
-}
-
 /**
  * Build a callback URL by appending selected state parameters.
  *
@@ -94,17 +77,11 @@ function create_return_to()
  */
 function create_callback_url($url)
 {
-    $state = [];
-    $return_to = create_return_to();
+    $state = create_state(['cat', 'code', 'test']);
+
+    $return_to = create_return_to($_SERVER['HTTP_REFERER']);
     if (!empty($return_to)) {
         $state['return_to'] = $return_to;
-    }
-
-    foreach (['cat', 'code', 'test'] as $key) {
-        $da = filter_input(INPUT_GET, $key, FILTER_SANITIZE_STRING);
-        if (!empty($da)) {
-            $state[$key] = $da;
-        }
     }
 
     if (!empty($state)) {
